more docs+move functions to parsers.go
